backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
alembic
python-jose[cryptography]
email-validator
pydantic-settings
scikit-learn
joblib
pandas
numpy
